scripts/requirements.txt
networkx
chromadb
sentence-transformers
ollama
gitpython
python-dotenv
tqdm
colorama
